requirements.txt
torch
torchvision
pyyaml
albumentations
timm